renamed authbackend folder to pascal case
