+++ api/models.py
@@ -3,7 +3,7 @@ from django.db import models
 from api.logistics.choice_enums import YEAR_IN_SCHOOL_CHOICES, LEARNING_GOALS_CHOICES, EXPERIENTIAL_LEARNING_HOURS_TYPES
 from api.reliability.validators import no_future_dates, hour_instance_validator, minutes_validator, student_id_validator
 
-from authBackEnd.apps.common import models as common_models
+from auth_backend.modules.common import models as common_models
 
 
 class TimeMaster(models.Model):
@@ -15,7 +15,6 @@ class TimeMaster(models.Model):
 
 
 class HourInstance(common_models.TimeStamp):
-    ##enharited the timestamp from authBackEnd
     """
     An instance of a single hour submission. Connected to its related student through the time master foreign key.
     """
