+++ auth_backend/modules/common/apps.py
@@ -2,4 +2,4 @@ from django.apps import AppConfig
 
 
 class CommonConfig(AppConfig):
-    name = 'apps.common'
+    name = 'modules.common'
