+++ auth_backend/modules/dashboard/apps.py
@@ -2,4 +2,4 @@ from django.apps import AppConfig
 
 
 class DashboardConfig(AppConfig):
-    name = 'apps.dashboard'
+    name = 'modules.dashboard'
