+++ auth_backend/modules/superAdmin/apps.py
@@ -1,5 +1,6 @@
 from django.apps import AppConfig
 
-#all the admin stuff
+
+# all the admin stuff
 class AdminConfig(AppConfig):
-    name = 'apps.superAdmin'
+    name = 'modules.superAdmin'
